refactor: Update quotationWithAllDataDto and createQuotationWithProducts models
The code changes in quotationWithAllDataDto.ts, createQuotationWithProducts.ts, and updateQuotationWithProducts.ts modify the models related to quotations and products. The "quotation" property has been renamed to "product" and the type has been updated to OmitCreateQuotationProductsDto[]. Additionally, a new "product" property has been added to the quotationWithAllDataDto model. These changes improve the clarity and consistency of the models.

This commit message follows the established convention of starting with a verb in the imperative form ("Update quotationWithAllDataDto and createQuotationWithProducts models") and provides a clear and concise summary of the changes made.
